fix: clean the main script and load FEDS from separate file
